Generate fed templates (#208)
* WIP

* Add beginnings of RIOT Makefile

* Implement LFC templates for RIOT (#210)

* Implement LFC templates for RIOT

* Remove makefile from coap-federated example

* Refactor string formatting a little

* Fix RIOT makefile generation

---------


* Add `generate.sh` file to generated fed templates (#215)

* Generate generate.sh files to invoke lfc from a federate

* Spotless

---------


* generate.sh -> run_lfc.sh

* Split generated CMake into Include.cmake and CMakeLists.txt

* Fix shell script

* Sanitize input to lf_build_generated_code

* Updates

* Fix riot test

* Fix riot example

* Run formatter

* FIx merge

* Debug CI

* correct path to enc

* Formatting

* Test gen-fed-templates in CI

* CI

* CI

* Dont do RIOT yet because of CI issue

* CI

* Formating

* Delete mistakenly committed files

---------

+++ src/federated.c
@@ -92,8 +92,8 @@ void FederatedOutputConnection_cleanup(Trigger *trigger) {
     tagged_msg->tag.microstep = sched->current_tag(sched).microstep;
 
     assert(self->bundle->serialize_hooks[self->conn_id]);
-    ssize_t msg_size = (*self->bundle->serialize_hooks[self->conn_id])(
-        self->staged_payload_ptr, self->payload_pool.size, tagged_msg->payload.bytes);
+    int msg_size = (*self->bundle->serialize_hooks[self->conn_id])(self->staged_payload_ptr, self->payload_pool.size,
+                                                                   tagged_msg->payload.bytes);
     if (msg_size < 0) {
       LF_ERR(FED, "Failed to serialize payload for federated output connection %p", trigger);
     } else {
